% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{dapproxfun}
\alias{dapproxfun}
\title{Wrapper of \code{stats::approxfun} to generate densities linear interpolation functions}
\usage{
dapproxfun(...)
}
\arguments{
\item{...}{Any parameters required by \code{\link[stats]{approxfun}} function to specify the interpolation.
Typically, this includes the x, y coordinates of the data points.}
}
\value{
returns a function performing (linear or constant) interpolation of the given data points. For a given
   set of x values, this function will return the corresponding interpolated values. It uses data stored in its
   environment when it was created.
}
\description{
\code{dapproxfun} applies \code{\link[stats]{approxfun}} to interpolate with splines, but it limits the output
   values to be positive to match the behavior of the density of probabilities function.
}
