% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{dsplinefun}
\alias{dsplinefun}
\title{Wrapper of \code{stats::splinefun} to generate densities and distribution spline interpolation functions}
\usage{
dsplinefun(...)
}
\arguments{
\item{...}{Any parameters required by \code{\link[stats]{splinefun}} function to specify the spline interpolation.
Typically, this includes the x, y coordinates of the data points.}
}
\value{
returns a function with formal arguments x and deriv, the latter defaulting to zero. This function can be
   used to evaluate the interpolating cubic spline (deriv = 0), or its derivatives (deriv = 1, 2, 3) at the
   points x, where the spline function interpolates the data points originally specified.
}
\description{
\code{dsplinefun} applies \code{\link[stats]{splinefun}} to generate an interpolation function with
splines, but it limits the output of the resulting interpolation function to
values between 0 and 1 to match the behavior of the density and distribution of probabilities functions.
}
