% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{ffunCreate}
\alias{ffunCreate}
\title{Create density or distribution of probabilities functions from data}
\usage{
ffunCreate(x.data, aprxf = dapproxfun, accumulate = FALSE, ...)
}
\arguments{
\item{x.data}{A numeric vector containing the data.}

\item{aprxf}{The creator of the interpolation function.}

\item{accumulate}{If \code{TRUE}, it indicates that the resulting function is a distribution function, i.e.,
an accumulated density function.}

\item{...}{All the other, optional, parameters required by \code{\link[stats]{density}} function to
generate a kernel density estimation from the data.}
}
\value{
Either a denstity or a distribution function from the data. If the parameter \code{accumulate} is
   TRUE, the resulting function will be an accumulated density of probabilities function, also known as
   a distribution of probabilities function, otherwise, the resulting function will be a density of
   probabilities function.

   To perform its task, \code{ffunCreate} generates a series of points in the domain and then interpolates
   using the interpolation function resulting from applying the given \code{aprxf} approximation function.
}
\description{
\code{ffunCreate} creates a density or a distribution of probabilites function from a given set of
   data. The resulting function will interpolate the intermediate values using the given \code{aprxf}
   approximation function.
}
\examples{
X <- log(rgamma(150,5)) # Random data generation
dfun <- ffunCreate(X)   # To create a density function from X
dfun(c(0.45, 1.84, 2.3))
plot(dfun, xlim=c(-0.1, 3.0))

pfun <- ffunCreate(X, dsplinefun, accumulate=T) # To create a distribution spline interpolated function
plot(pfun, xlim=c(-0.1, 3.0))

}
\seealso{
\code{\link[stats]{density}}
}
