% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{pSfunCreate}
\alias{pSfunCreate}
\title{Create density of probabilities spline interpolated functions from data}
\usage{
pSfunCreate(x.data, ...)
}
\arguments{
\item{x.data}{A numeric vector containing the data.}

\item{...}{All the other, optional, parameters required by \code{\link[stats]{density}} function to
generate a kernel density estimation from the data.}
}
\value{
returns a function performing spline interpolation on a set of points generated from a kernel
   density of probabilities from the given \code{x.data}. In this case, the resulting kernel desity
   points are accumulated prior to the interpolation.
}
\description{
\code{pSfunCreate} takes a set of data and generates a distribution of probabilities function using
   splines to interpolate between the resulting sample points. A distribution of probabilities is
   cumulative distribution function.
}
